Localized block registration
